// src/utils/types/interfaceCreatedModal.ts
export interface CreatedModalProps {
    sizeIcon: number;
    colorIcon: string;
    color: string;
    apresentationIcon: string;
    firstIconName: string;
    labelIconFirst: string;
    firstAction: () => void;
    apresentationColor?: string;
    secondIconName?: string;
    labelIconSecond?: string;
    secondAction?: () => void;
  }